Major refactoring: renamed classes, changed ints to doubles, added update methods for nodes

+++ src/application/Actions/AbstractAction.java
@@ -17,7 +17,7 @@ public abstract class AbstractAction {
 	
 	public AbstractTurtle myDrawer;
 	public Point2D nextPoint;
-	public int myValue;
+	public double myValue;
 	/**
 	 * update takes in an AbstractDrawer and updates it according to the specifications of this type of Action.
 	 * @param currentDrawer
@@ -25,7 +25,7 @@ public abstract class AbstractAction {
 	public abstract void update(Turtle turt, SLogoCanvas canvas);
 	public abstract String toString();
 	
-	public int getValue() {
+	public double getValue() {
 		return myValue;
 	}
 }
